float result

+++ enginetest/join_op_tests.go
@@ -185,7 +185,7 @@ SELECT SUM(x) FROM xy WHERE x IN (
     )
   ) AND
   x = 2;`,
-				Expected: []sql.Row{{2}},
+				Expected: []sql.Row{{float64(2)}},
 			},
 			{
 				Query:    "select * from ab left join uv on a = u where exists (select * from uv where false)",
